DM-45522: Docs for direct Kafka connections
The upcoming app metrics events functionality will require direct
connection to Kafka in order for apps to emit metrics events. There are
helpers for making this connection in the app code
[here](https://github.com/lsst-sqre/safir/pull/302), but the Kafka auth
credentials need to be generated first. This PR adds documentation for
how to generate those credentials for the Sasquatch Kafka clusters.